refactor: rename webapp to chemistry_calculators
- Rename apps/webapp, templates/webapp, tests/webapp directories
- Update INSTALLED_APPS, URL includes, app config class
- Update all template extends/include references (12 files)
- Update test imports and CI/CD workflow
- Update README.md and OPS_PLAYBOOK.md documentation
