+++ apps/chemistry_calculators/calculations/equilibria.py
