+++ templates/chemistry_calculators/calculator/_calculator_base.html
@@ -1,4 +1,4 @@
-{% extends 'webapp/base.html' %}
+{% extends 'chemistry_calculators/base.html' %}
 
 {% block main %}
 <div class="my-3">
