+++ templates/chemistry_calculators/calculator/_tag_input_field.html
@@ -4,7 +4,7 @@ Renders a container that the SubstanceTagInput JS widget transforms into
 a tag/chip input with autocomplete.
 
 Usage:
-  {% include 'webapp/calculator/_tag_input_field.html' with field=field %}
+  {% include 'chemistry_calculators/calculator/_tag_input_field.html' with field=field %}
 {% endcomment %}
 <div class="mb-4">
   <label for="fallback-{{ field.name }}" class="form-label">{{ field.label }}</label>
