+++ templates/chemistry_calculators/calculator/dilution.html
@@ -1,4 +1,4 @@
-{% extends 'webapp/calculator/_calculator_base.html' %}
+{% extends 'chemistry_calculators/calculator/_calculator_base.html' %}
 
 {% block calculator_title %}Dilution Calculator{% endblock %}
 
@@ -16,7 +16,7 @@
       </div>
     {% endif %}
     {% for item in dilution_fields %}
-      {% include 'webapp/calculator/_dilution_pair.html' with item=item %}
+      {% include 'chemistry_calculators/calculator/_dilution_pair.html' with item=item %}
     {% endfor %}
   </form>
 {% endblock %}
