+++ templates/chemistry_calculators/calculator/equilibria.html
@@ -1,4 +1,4 @@
-{% extends 'webapp/calculator/_calculator_base.html' %}
+{% extends 'chemistry_calculators/calculator/_calculator_base.html' %}
 {% load static %}
 
 {% block calculator_title %}Chemical Equilibria Calculator{% endblock %}
