+++ templates/chemistry_calculators/components/_result_card.html
@@ -2,7 +2,7 @@
 Reusable result card component for calculator pages.
 
 Usage:
-  {% include 'webapp/components/_result_card.html' with result_id="mw-result" result_content_block=calculator_result %}
+  {% include 'chemistry_calculators/components/_result_card.html' with result_id="mw-result" result_content_block=calculator_result %}
 
 Variables:
   result_id       — ID for the result container (used by clipboard copy)
