+++ templates/chemistry_calculators/landing.html
@@ -1,4 +1,4 @@
-{% extends 'webapp/base.html' %}
+{% extends 'chemistry_calculators/base.html' %}
 
 {% block header %}
