+++ templates/hplc_simulator/hplc_base.html
@@ -1,4 +1,4 @@
-{% extends "webapp/base.html" %}
+{% extends "chemistry_calculators/base.html" %}
 {% load static %}
 
 {% block extra_head %}
